Adding columns to dgv Rewritten Load_Data

+++ ResidualMaterials/Balance.cs
@@ -4,9 +4,9 @@
     {
         public int BalanceID { get; set; }
         public bool Type { get; set; }
-        public double Dim { get; set; }
-        public double Length { get; set; }
-        public double W { get; set; }
-        public double H { get; set; }
+        public decimal Dim { get; set; }
+        public decimal Length { get; set; }
+        public decimal W { get; set; }
+        public decimal H { get; set; }
     }
 }
